feat(vsc+core): remove bundled sample config and use offline page
