+++ packages/fx-core/tests/common/samples-config-v3.json
