+++ packages/vscode-extension/src/controls/sampleGallery/SampleGallery.scss
@@ -1,0 +1,64 @@
+.sample-gallery {
+  margin: 80px 100px;
+
+  .section {
+    display: flex;
+    margin-bottom: 36px;
+  }
+
+  #title {
+    margin-left: auto;
+    margin-right: auto;
+  }
+
+  .logo {
+    width: 28px;
+    height: 28px;
+    font-size: 28px;
+    color: var(--vscode-textLink-foreground);
+  }
+
+  .title {
+    width: auto;
+    margin-left: 12px;
+
+    h1 {
+      font-size: 26px;
+      line-height: 31px;
+      color: var(--vscode-editor-foreground);
+      margin-top: 0px;
+      margin-bottom: 0px;
+      font-weight: 600;
+    }
+
+    h3 {
+      font-size: 12px;
+      line-height: 14px;
+      color: var(--vscode-editor-foreground);
+      margin-top: 0px;
+      margin-bottom: 0px;
+      font-weight: normal;
+    }
+  }
+
+  .sample-stack {
+    display: grid;
+    max-width: 1164px;
+    margin: 0 auto;
+    grid-template-columns: repeat(12, 1fr);
+    grid-template-rows: auto;
+    grid-gap: 20px;
+  }
+}
+
+@media only screen and (max-width: 640px){
+  .sample-gallery{
+    margin: 100px 32px;
+  }
+}
+
+@media only screen and (max-width: 480px) {
+  .sample-gallery {
+    min-width: 320px;
+  }
+}
