+++ packages/vscode-extension/src/controls/sampleGallery/offlinePage.scss
@@ -1,0 +1,26 @@
+.offlinePage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  
+  .offlineImage {
+    margin-top: 120px;
+  }
+
+  .offlineTitle {
+    text-align: center;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 22px;
+    margin: 6px;
+  }
+
+  .offlineMessage {
+    text-align: center;
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 20px;
+  }
+}
