+++ packages/vscode-extension/src/controls/sampleGallery/sampleCard.scss
@@ -1,65 +1,18 @@
 .sample-gallery {
-  margin: 80px 100px;
 
-  .section {
-    display: flex;
-    margin-bottom: 36px;
-  }
-
-  #title {
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  .logo {
-    width: 28px;
-    height: 28px;
-    font-size: 28px;
-    color: var(--vscode-textLink-foreground);
-  }
-
-  .title {
-    width: auto;
-    margin-left: 12px;
-
-    h1 {
-      font-size: 26px;
-      line-height: 31px;
-      color: var(--vscode-editor-foreground);
-      margin-top: 0px;
-      margin-bottom: 0px;
-      font-weight: 600;
-    }
-
-    h3 {
-      font-size: 12px;
-      line-height: 14px;
-      color: var(--vscode-editor-foreground);
-      margin-top: 0px;
-      margin-bottom: 0px;
-      font-weight: normal;
-    }
-  }
-
-  .sample-stack {
-    display: grid;
-    max-width: 1164px;
-    margin: 0 auto;
-    grid-template-columns: repeat(12, 1fr);
-    grid-template-rows: auto;
-    grid-gap: 20px;
-  }
-
-  .box1, .box5, .box9, .box13, .box17, .box21 {
+  .sample-stack .sample-card:nth-child(4n + 1) {
     grid-column: 1/4;
   }
-  .box2, .box6, .box10, .box14, .box18, .box22 {
+
+  .sample-stack .sample-card:nth-child(4n + 2) {
     grid-column: 4/7;
   }
-  .box3, .box7, .box11, .box15, .box19, .box23 {
+
+  .sample-stack .sample-card:nth-child(4n + 3) {
     grid-column: 7/10;
   }
-  .box4, .box8, .box12, .box16, .box20, .box24 {
+
+  .sample-stack .sample-card:nth-child(4n + 4) {
     grid-column: 10/13;
   }
 
@@ -79,7 +32,7 @@
       flex-wrap: wrap;
     }
 
-    .triangle{
+    .triangle {
       width: 0;
       height: 0;
       border-style: solid;
@@ -92,22 +45,22 @@
       z-index: 2;
     }
 
-    .star{
-        display: block;
-        text-align: center;
-        height: 12px;
-        width: 12px;
-        position: absolute;
-        margin: 2px;
-        top: -16px;
-        left: -16px;
-        right: 0;
-        bottom: 0;
-        z-index: 3;
-        color: #CCCCCC;
+    .star {
+      display: block;
+      text-align: center;
+      height: 12px;
+      width: 12px;
+      position: absolute;
+      margin: 2px;
+      top: -16px;
+      left: -16px;
+      right: 0;
+      bottom: 0;
+      z-index: 3;
+      color: #CCCCCC;
     }
 
-    .ms-Image{
+    .ms-Image {
       border-radius: 6px 6px 0 0;
     }
 
@@ -115,7 +68,7 @@
       margin: 8px 0px 0px 8px;
     }
 
-    .estimation-time{
+    .estimation-time {
       padding-left: 8px;
       height: 24px;
       color: var(--vscode-editor-foreground);
@@ -124,26 +77,26 @@
       margin-top: 8px;
       margin-bottom: 2px;
 
-      .watch{
-          width: 16px;
-          height: 16px;
-          margin-top: auto;
-          margin-bottom: auto;
-      }
-
-      .ms-Image{
+      .watch {
+        width: 16px;
+        height: 16px;
         margin-top: auto;
         margin-bottom: auto;
       }
 
-      label{
+      .ms-Image {
+        margin-top: auto;
+        margin-bottom: auto;
+      }
+
+      label {
         margin-top: auto;
         margin-bottom: auto;
         cursor: pointer;
       }
     }
 
-    .configuration{
+    .configuration {
       padding-left: 8px;
       height: auto;
       color: var(--vscode-editor-foreground);
@@ -151,19 +104,19 @@
       font-size: 11px;
       height: 24px;
 
-      .setting{
-          width: 16px;
-          height: 16px;
-          margin-top: auto;
-          margin-bottom: auto;
-      }
-
-      .ms-Image{
+      .setting {
+        width: 16px;
+        height: 16px;
         margin-top: auto;
         margin-bottom: auto;
       }
 
-      label{
+      .ms-Image {
+        margin-top: auto;
+        margin-bottom: auto;
+      }
+
+      label {
         width: 242px;
         margin-top: auto;
         margin-bottom: auto;
@@ -209,45 +162,43 @@
 }
 
 @media only screen and (max-width: 1024px) {
-  .sample-gallery{
-    .box1, .box4, .box7, .box10, .box13, .box16, .box19, .box22 {
+  .sample-gallery {
+
+    .sample-stack .sample-card:nth-child(3n + 1) {
       grid-column: 1/5;
     }
-    .box2, .box5, .box8, .box11, .box14, .box17, .box20, .box23 {
+
+    .sample-stack .sample-card:nth-child(3n + 2) {
       grid-column: 5/9;
     }
 
-    .box3, .box6, .box9, .box12, .box15, .box18, .box21, .box24 {
+    .sample-stack .sample-card:nth-child(3n + 3) {
       grid-column: 9/13;
     }
   }
 }
 
-@media only screen and (max-width: 813px){
-  .sample-gallery{
-    .box1, .box3, .box5, .box7, .box9, .box11, .box13, .box15, .box17, .box19, .box21, .box23 {
+@media only screen and (max-width: 813px) {
+  .sample-gallery {
+
+    .sample-stack .sample-card:nth-child(2n + 1) {
       grid-column: 1/7;
     }
-    .box2, .box4,.box6,.box8, .box10,.box12, .box14, .box16, .box18, .box20, .box22, .box24 {
+
+    .sample-stack .sample-card:nth-child(2n + 2) {
       grid-column: 7/13;
     }
   }
 }
 
-@media only screen and (max-width: 640px){
-  .sample-gallery{
-    margin: 100px 32px;
-  }
-}
-
 @media only screen and (max-width: 480px) {
   .sample-gallery {
-    min-width: 320px;
 
-  .box1, .box2, .box3, .box4, .box5, .box6, .box7, .box8, .box9, .box10, .box11, .box12, .box13, .box14, .box15, .box16, .box17, .box18, .box19, .box20, .box21, .box22, .box23, .box24 {
+    .sample-stack > .sample-card {
       grid-column: 1/13;
     }
   }
+
   .sample-card {
     justify-self: center;
     grid-column: 1/5;
