+++ packages/vscode-extension/src/controls/sampleGallery/sampleDetailPage.scss
