+++ packages/vscode-extension/webpack.config.js
@@ -59,10 +59,15 @@ const config = {
         use: ["style-loader", "css-loader"],
       },
       {
-        test: /\.(jpg|png|svg|gif)$/,
+        test: /\.(jpg|png|gif)$/,
         use: {
           loader: "url-loader",
         },
+      },
+      {
+        test: /\.svg$/i,
+        issuer: /\.tsx?$/,
+        use: ["@svgr/webpack"],
       },
       {
         test: /node_modules[\\|/](yaml-language-server|vscode-languageserver|vscode-json-languageservice|prettier)/,
